✨ TimeUtils (support for ticks in durations)
