+++ src/test/resources/literals/Duration.txt
@@ -7,5 +7,7 @@ test:
 	assert "%{dur}%" = "a day and 6 minutes" with "{dur} should be 'a day and 6 minutes': %{dur}%"
 
 	assert 5.5 minutes is equal to 5 minutes and 30 seconds with "{dur} should equal 5 minutes and 30 seconds: %{dur}%"
+	assert 3 ticks = 0.15 seconds with "3 ticks should equal 0.15 seconds"
+	assert a tick = 50 milliseconds with "a tick should equal 50 milliseconds"
 
 	throws 5 days, 3 minutes, and 5 seconds
